+++ telepot/namedtuple.py
@@ -126,6 +126,8 @@ Chat = _create_class('Chat', [
            'description',
            'invite_link',
            _Field('pinned_message', constructor=_Message),
+           'sticker_set_name',
+           'can_set_sticker_set',
        ])
 
 # incoming
@@ -449,6 +451,7 @@ Message = _create_class('Message', [
               'author_signature',
               'text',
               _Field('entities', constructor=MessageEntityArray),
+              _Field('caption_entities', constructor=MessageEntityArray),
               _Field('audio', constructor=Audio),
               _Field('document', constructor=Document),
               _Field('game', constructor=Game),
@@ -542,6 +545,7 @@ InputTextMessageContent = _create_class('InputTextMessageContent', [
 InputLocationMessageContent = _create_class('InputLocationMessageContent', [
                                   'latitude',
                                   'longitude',
+                                  'live_period',
                               ])
 
 # outgoing
@@ -685,6 +689,7 @@ InlineQueryResultLocation = _create_class('InlineQueryResultLocation', [
                                 'latitude',
                                 'longitude',
                                 'title',
+                                'live_period',
                                 'reply_markup',
                                 'input_message_content',
                                 'thumb_url',
@@ -817,3 +822,26 @@ InlineQueryResultCachedAudio = _create_class('InlineQueryResultCachedAudio', [
                                    'reply_markup',
                                    'input_message_content',
                                ])
+
+# outgoing
+InputMediaPhoto = _create_class('InputMediaPhoto', [
+                      _Field('type', default='photo'),
+                      'media',
+                      'caption',
+                  ])
+
+# outgoing
+InputMediaVideo = _create_class('InputMediaVideo', [
+                      _Field('type', default='video'),
+                      'media',
+                      'caption',
+                      'width',
+                      'height',
+                      'duration',
+                  ])
+
+# incoming
+ResponseParameters = _create_class('ResponseParameters', [
+                         'migrate_to_chat_id',
+                         'retry_after',
+                     ])